Fix kube-rbac-proxy probe TLS errors (#24)
Although the TCP {liveness,readiness}Probes of the kube-rbac-proxy
container work as expected, i.e. succeed because a connection can
be established, kube-rbac-proxy logs an error because of the TLS
failures.

We can skip logging all these errors by replacing the TCP probes with
ExecActions ones, which use `curl` to try to fetch the headers. This
way we verify that the kube-rbac-proxy is up and at the same time we
ignore any HTTP 401 errors, to skip making an authenticated request.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>