Add RBAC manifests in CSV when generating the bundle
When packaging additional manifests alongside the CSV, OLM will not
perform any templating on those. Currently, additional RBAC manifests are
packaged alongside the HabanaAI operator CSV, i.e. for the node-labeler,
node-metrics, driver and device-plugin workloads. The RoleBindings of
those ServiceAccounts include the namespace of the ServiceAccount, which
should be the operator's namespace. The latter can be chosen dynamically
by the user when installing the operator. But since OLM won't perform any
templating on those RoleBindings, the namespace under the subjects spec
of the latter is now hardcoded to `habana-ai-operator`.

To resolve this issue, we inject the necessary RBAC manifests directly
in the CSV via the `--extra-service-accounts` argument of the
operator-sdk command when creating the OLM bundle. This way the RBAC
manifests are created and templated properly by OLM when parsing and
applying the resources included in the CSV.

- https://sdk.operatorframework.io/docs/building-operators/golang/operator-scope/#configuring-watch-namespaces-dynamically
- https://sdk.operatorframework.io/docs/advanced-topics/multi-sa/#updating-the-makefile-to-use---extra-service-accounts

Signed-off-by: Michail Resvanis <mresvani@redhat.com>