Replace additional ClusterRoleBindings to RoleBindings
Since the operator is installed via OLM in its own namespace and the
additional RBAC manifests needed for the HabanaAI workload are also
applied in that same namespace, there is no need for ClusterRoleBindings
for those ServiceAccounts. RoleBindings can be used instead to provide
the same permissions, but only in that specific namespace and not
cluster-wide.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>